FPL-227: Fixed issue with show hide on EPO Reason
Callback has been revert to an aboutToStartCallback on the enter grounds event. The EPO_REASONING_SHOW field is now defined as a multiSelectList and populated with an array of strings by the callback event. FieldShowCondition in caseEventToField has been refactored to CONTAINS instead of equals. This appears to have fixed errors around fields disappearing and not displaying correctly