Remove e2e tests from PR pipeline for FPL (#4566)
* Remove e2e tests from PR pipeline for FPL

The end to end tests are overkill. They only assert
UI behaviour which is owned by XUI. Running it against every
PR is too much. These should be running in the nightly pipeline only.

* Remove as pipeline works with it commented out.

As mentioned in previous commit. These are being removed as
UI tests assert XUI functionality and are overkill for our pipelines.

We can run them in a nightly pipeline instead of per PR.

* Comment out temp to investigate use of yarnBuilder in our Jenkins pipeline

* Remove commented code as it's unused and unneeded